requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.11.0=pypi_0
appdirs=1.4.4=py_0
astunparse=1.6.3=pypi_0
attrs=20.3.0=pyhd3eb1b0_0
black=19.10b0=py_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=pypi_0
certifi=2020.11.8=py38haa95532_0
click=7.1.2=py_0
gast=0.3.3=pypi_0
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.34.0=pypi_0
h5py=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
markdown=3.3.3=pypi_0
mypy_extensions=0.4.3=py38_0
numpy=1.18.5=pypi_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1h=he774522_0
opt-einsum=3.3.0=pypi_0
pathspec=0.7.0=py_0
pip=20.3=py38haa95532_0
protobuf=3.14.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygame=2.0.0=pypi_0
python=3.8.5=h5fd99cc_1
regex=2020.11.13=py38h2bbff1b_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
setuptools=50.3.2=py38haa95532_2
six=1.15.0=pypi_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.4.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow-gpu=2.3.1=pypi_0
tensorflow-gpu-estimator=2.3.0=pypi_0
termcolor=1.1.0=pypi_0
toml=0.10.1=py_0
typed-ast=1.4.1=py38he774522_0
typing_extensions=3.7.4.3=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
werkzeug=1.0.1=pypi_0
wheel=0.36.0=pyhd3eb1b0_0
wincertstore=0.2=py38_0
wrapt=1.12.1=pypi_0
zlib=1.2.11=h62dcd97_4
